Redirecting the user to the activity page
