+++ ui/components/Wallet/WalletAssetList.tsx
@@ -6,12 +6,11 @@ import {
   SwappableAsset,
 } from "@tallyho/tally-background/assets"
 import WalletAssetListItem from "./WalletAssetListItem"
-import AssetWarningSlideUp from "./UnverifiedAsset/AssetWarningSlideUp"
+import AssetWarningWrapper from "./UnverifiedAsset/AssetWarningWrapper"
 
 type WalletAssetListProps = {
   assetAmounts: CompleteAssetAmount<SwappableAsset>[]
   initializationLoadingTimeExpired: boolean
-  onCloseWarningSlideUp?: () => void
 }
 
 export default function WalletAssetList(
@@ -21,11 +20,7 @@ export default function WalletAssetList(
     keyPrefix: "wallet.activities",
   })
 
-  const {
-    assetAmounts,
-    initializationLoadingTimeExpired,
-    onCloseWarningSlideUp,
-  } = props
+  const { assetAmounts, initializationLoadingTimeExpired } = props
 
   const [warnedAsset, setWarnedAsset] = useState<
     CompleteAssetAmount<SmartContractFungibleAsset>["asset"] | null
@@ -35,17 +30,12 @@ export default function WalletAssetList(
 
   return (
     <>
-      {warnedAsset && (
-        <AssetWarningSlideUp
-          asset={warnedAsset}
-          close={() => {
-            setWarnedAsset(null)
-            if (onCloseWarningSlideUp) {
-              onCloseWarningSlideUp()
-            }
-          }}
-        />
-      )}
+      <AssetWarningWrapper
+        asset={warnedAsset}
+        close={() => {
+          setWarnedAsset(null)
+        }}
+      />
       <ul>
         {assetAmounts.map((assetAmount) => (
           <WalletAssetListItem
