+++ ui/pages/SingleAsset.tsx
@@ -30,9 +30,9 @@ import WalletActivityList from "../components/Wallet/WalletActivityList"
 import SharedBackButton from "../components/Shared/SharedBackButton"
 import SharedTooltip from "../components/Shared/SharedTooltip"
 import { blockExplorer } from "../utils/constants"
-import AssetWarningSlideUp from "../components/Wallet/UnverifiedAsset/AssetWarningSlideUp"
 import AssetVerifyToggler from "../components/Wallet/UnverifiedAsset/AssetVerifyToggler"
 import { trimWithEllipsis } from "../utils/textUtils"
+import AssetWarningWrapper from "../components/Wallet/UnverifiedAsset/AssetWarningWrapper"
 
 const MAX_SYMBOL_LENGTH = 10
 
@@ -109,14 +109,12 @@ export default function SingleAsset(): ReactElement {
 
   return (
     <>
-      {warnedAsset && (
-        <AssetWarningSlideUp
-          asset={warnedAsset}
-          close={() => {
-            setWarnedAsset(null)
-          }}
-        />
-      )}
+      <AssetWarningWrapper
+        asset={warnedAsset}
+        close={() => {
+          setWarnedAsset(null)
+        }}
+      />
       <div className="navigation standard_width_padded">
         <SharedBackButton path="/" />
         {isEnabled(FeatureFlags.SUPPORT_UNVERIFIED_ASSET) && (
